Extract endpoint multipart zip helpers
